Bugfix for Safari and proper minification
Because HTML imports in Safari are asynchronous there is a race condition that the store is undefined. To fix this the redux scripts were moved out of the elements.html  into the index.html.
A small bugfix in the gulpfile.js ensures that the HTML files are minified.
Some small bugfixes in the cloud page